# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/operations": {
            "get": {
                "description": "retrieve the list of operations",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Operations"
                ],
                "summary": "Get the operations list",
                "operationId": "get-operations",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter parameter",
                        "name": "filter_param",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter value",
                        "name": "filter_value",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort field",
                        "name": "sort_field",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort order",
                        "name": "sort_order",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/repositories.PaginatedOperations"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorMessage"
                        }
                    }
                }
            },
            "post": {
                "description": "create a new operation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Operations"
                ],
                "summary": "Create a new operation",
                "operationId": "post-operation",
                "parameters": [
                    {
                        "description": "Operation object",
                        "name": "operation",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.OperationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.OperationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorMessage"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorMessage"
                        }
                    }
                }
            }
        },
...
        "/api/v1/operations/{id}": {
            "get": {
                "description": "retrieve an operation by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Operations"
                ],
                "summary": "Get an operation",
                "operationId": "get-operation-by-id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Operation ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/operation.OperationWithLogs"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorMessage"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorMessage"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "operation.OperationWithLogs": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "string"
                },
                "blockchain_status": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "domain": {
                    "type": "string"
                },
                "fireblocks_id": {
                    "type": "string"
                },
                "fireblocks_status": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "logs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/repositories.OperationLog"
                    }
                },
                "operator": {
                    "type": "string"
                },
                "transaction_hash": {
                    "type": "string"
                },
                "transaction_link": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "repositories.Operation": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "string"
                },
                "blockchain_status": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "domain": {
                    "type": "string"
                },
                "fireblocks_id": {
                    "type": "string"
                },
                "fireblocks_status": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "operator": {
                    "type": "string"
                },
                "transaction_hash": {
                    "type": "string"
                },
                "transaction_link": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "repositories.OperationLog": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "error": {},
                "fireblocks_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "operation_id": {
                    "type": "string"
                },
                "payload": {},
                "response": {}
            }
        },
        "repositories.PaginatedOperations": {
            "type": "object",
            "properties": {
                "current_page": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/repositories.Operation"
                    }
                },
                "next_page": {
                    "type": "integer"
                },
                "previous_page": {
                    "type": "integer"
                },
                "total_count": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
        "types.ErrorMessage": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "types.OperationRequest": {
            "type": "object",
            "required": [
                "amount",
                "blockchain_id",
                "domain",
                "operator",
                "token_id",
                "type"
            ],
            "properties": {
                "amount": {
                    "type": "string",
                    "example": "2.75"
                },
                "blockchain_id": {
                    "type": "string",
                    "example": "66f6fe7eccc6398d39e981f9"
                },
                "domain": {
                    "type": "string",
                    "enum": [
                        "GET-BRAZA",
                        "BRAZA-ON",
                        "BRAZA-DESK"
                    ],
                    "example": "GET-BRAZA"
                },
                "operator": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                },
                "token_id": {
                    "type": "string",
                    "example": "66f74acbba6b56108cb3e80a"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "MINT",
                        "BURN"
                    ],
                    "example": "MINT,BURN"
                }
            }
        },
        "types.OperationResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "operation was accepted to be processed on blockchain"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
    }
}